fix: Test-Static lint failures + e2e coverage for warp/gun/tubes bars
The Prettier formatting (warp.ts) and import-sort (full-system-status.ts)
issues from the previous commit were failing the Test-Static CI check;
fixed both. Also extends e2e coverage of the .sw-bar treatment to the
warp, chain gun, and tubes-status panels via a shared
expectNonInteractiveBar() helper, mirroring the addBarBlade call sites.

🤖 Automated by Claude Code
